Complete COPA and add tests
