+++ docs/ResponsibleAI.md
@@ -1,0 +1,39 @@
+# Responsible Artificial Intelligence at ULOSINO
+
+ULOSINO has been built using contemporary development techniques and industry practices.
+
+This approach involves the use of artificial intelligence (AI) and machine learning (ML) technologies.
+
+Project contributors should review code generated by AI.
+
+ULOSINO will never deploy AI technologies that don't emphasise developer privacy and repository safety. For example, we won't deploy automatic code refactoring on a CI pipeline that automatically updates when passed (without human interaction or intervention).
+
+## Technology-Specific Details
+
+### GitHub Copilot
+
+[Noah Stanley (@fernpolo)](https://github.com/fernpolo), the primary contributor, is known to have access to the GitHub Copilot Technical Preview.
+
+Some Source Code present across this repository may have been generated by GitHub Copilot. GitHub Copilot is a generation tool; the code generated by the system remains copyright of the contributor and subject to the Mozilla Public License 2.0.
+
+Code generated by GitHub Copilot may not be accurate and the code may require testing.
+
+[Learn More about GitHub Copilot...](https://copilot.github.com)
+
+### CodeQL Security Scanning
+
+Every line of Source Code is automatically reviewed by CodeQL technologies as part of our CI pipeline.
+
+CodeQL aims to detect security vulnerabilities in code. If vulnerabilities are found in code attempting to merge onto the `main` branch, the check will fail and merging may not be permitted.
+
+[Learn More about CodeQL Security Scanning on GitHub...](https://docs.github.com/en/code-security/code-scanning/automatically-scanning-your-code-for-vulnerabilities-and-errors/about-code-scanning-with-codeql)
+
+### GitHub Dependabot
+
+Dependabot checks for package dependency updates every day.
+
+It works with `npm` to check updates to `package.json` packages. If a newer version is available, Dependabot automatically opens a PR with the updated `package.json`. It can also open recommendation PRs if packages listed on `package.json` use versions used are known to contain security vulnerabilities.
+
+Dependabot PRs are subject to our standard pull request checks (CI/CD pipeline) and cannot be automatically merged.
+
+[Learn More about Dependabot on GitHub...](https://docs.github.com/en/code-security/dependabot)
